+++ lib/src/blocs/page/configuration_changed_event.dart
@@ -3,7 +3,7 @@ import 'event.dart';
 
 /// Emit this when configuration of the screen changes.
 ///
-/// The configuration will then be read with [PageBloc.getConfiguration].
+/// The configuration will then be read with [CPageBloc.getConfiguration].
 class PageBlocConfigurationChangedEvent extends PageBlocEvent {
   const PageBlocConfigurationChangedEvent();
 }