+++ lib/src/blocs/page_stack/event.dart
@@ -1,6 +1,6 @@
 import 'bloc.dart';
 
-/// The base class for [PageStackBloc] events.
+/// The base class for [CPageStackBloc] events.
 abstract class PageStackBlocEvent {
   const PageStackBlocEvent();
 }