+++ lib/src/blocs/page_stack/page_event.dart
@@ -4,15 +4,18 @@ import '../page/configuration.dart';
 import '../page/event.dart';
 import 'event.dart';
 
-class PageStackPageBlocEvent<C extends PageConfiguration, R>
+class CPageStackPageBlocEvent<C extends PageConfiguration, R>
     extends PageStackBlocEvent {
-  final AbstractPage<C, R> page;
-  final PageBloc<C, R>? bloc;
+  final CAbstractPage<C, R> page;
+  final CPageBloc<C, R>? bloc;
   final PageBlocEvent pageBlocEvent;
 
-  PageStackPageBlocEvent({
+  CPageStackPageBlocEvent({
     required this.page,
     required this.bloc,
     required this.pageBlocEvent,
   });
 }
+
+typedef PageStackPageBlocEvent<R>
+    = CPageStackPageBlocEvent<PageConfiguration, R>;