+++ lib/src/pages/stateless_material.dart
@@ -3,17 +3,19 @@ import 'package:flutter/widgets.dart';
 import '../blocs/page/configuration.dart';
 import 'material.dart';
 
-abstract class StatelessMaterialPage<C extends PageConfiguration, R>
-    extends AbstractMaterialPage<C, R> {
+abstract class CStatelessMaterialPage<C extends PageConfiguration, R>
+    extends CAbstractMaterialPage<C, R> {
   final C? configuration;
 
   @override
   C? getConfiguration() => configuration;
 
-  StatelessMaterialPage({
+  CStatelessMaterialPage({
     required ValueKey<String> super.key,
     required super.child,
     super.factoryKey,
     this.configuration,
   });
 }
+
+typedef StatelessMaterialPage<R> = CStatelessMaterialPage<PageConfiguration, R>;